Fix compilation errors and warnings - move header files to src, adjust const variables
