+++ .gitignore
@@ -1,3 +1,5 @@
+.DS_Store
+
 # Prerequisites
 *.d
 .DS_Store
