+++ src/callbacks.h
@@ -19,6 +19,6 @@ uint8_t GLOBAL_SCI_FIELD_NUM,
 void init_callbacks();
 void PAY_CMD_Tx_data_callback(uint8_t* data, uint8_t *len);
 void EPS_CMD_Tx_data_callback(uint8_t* data, uint8_t *len);
-void data_rx_mob_callback(uint8_t* data, uint8_t len);
+void data_rx_mob_callback(const uint8_t* data, uint8_t len);
 
 #endif
