+++ src/global_header.h
