+++ src/housekeeping.h
