+++ src/obc.h
@@ -1,24 +1,22 @@
 #ifndef OBC_H
 #define OBC_H
 
+#include <stdint.h>
 
-// standard C ;ibraries
-#include  	<stdbool.h>
-#include  	<stdint.h>
+#include <uart/uart.h>
+#include <uart/log.h>
+#include <timer/timer.h>
+#include <can/can.h>
 
-// avr-libc includes
-#include 	<avr/io.h>
-#include 	<util/delay.h>
+#include "can_queue.h"
+#include "cmd_queue.h"
+#include "callbacks.h"
 
-// lib-common includes
-#include 	<spi/spi.h>
-#include 	<uart/uart.h>
-#include 	<uart/log.h>
-#include 	<can/can.h>
+#define PAY_CMD_TX_MOB 3
+#define EPS_CMD_TX_MOB 4
+#define DATA_RX_MOB 5
 
-// project file includes
-#include 	"rtc.h"
-#include 	"mem.h"
-
+Can_queue sci_tx_queue, pay_hk_tx_queue, eps_hk_tx_queue;
+Cmd_queue obc_queue;
 
 #endif
